enhancement(content-manager): refactor MCP module into handlers, schemas, and utilities
### What does it do?

Breaks up the monolithic `derive-content-type-mcp-tools.ts` (~1900 lines) into focused modules:
- `handlers/collection-handlers.ts` and `handlers/single-type-handlers.ts` — CRUD tool handler logic per content-type kind
- `schemas/` — Zod schemas split by concern: blocks, data, filters, input, output, sort
- `permissions.ts` — locale/action permission resolution helpers
- `types.ts` — shared MCP type definitions
- `utils.ts` — shared utility functions
- `derive-content-type-mcp-tools.ts` now acts as a thin orchestrator

### Why is it needed?

The single file had grown to ~1900 lines, mixing schema definitions, permission logic, handler implementations, and type declarations. The split improves readability, testability, and makes it easier to extend individual concerns without touching unrelated code.

### How to test it?

Run the existing MCP unit tests:
```
yarn jest packages/core/content-manager/server/src/mcp
```
Verify MCP tools still register and behave correctly against a running Strapi instance with collection types and single types.
