+++ packages/core/content-manager/server/src/mcp/handlers/index.ts
@@ -1,0 +1,20 @@
+export {
+  createCollectionListHandler,
+  createCollectionGetHandler,
+  createCollectionCreateHandler,
+  createCollectionUpdateHandler,
+  createCollectionDeleteHandler,
+  createCollectionPublishHandler,
+  createCollectionUnpublishHandler,
+  createCollectionDiscardDraftHandler,
+} from './collection-handlers';
+
+export {
+  singleCreateOrUpdate,
+  createSingleGetHandler,
+  createSingleWriteHandler,
+  createSingleDeleteHandler,
+  createSinglePublishHandler,
+  createSingleUnpublishHandler,
+  createSingleDiscardDraftHandler,
+} from './single-type-handlers';
