+++ packages/core/content-manager/server/src/mcp/schemas/blocks-schema.ts
