+++ packages/core/content-manager/server/src/mcp/schemas/index.ts
@@ -1,0 +1,16 @@
+export { buildBlocksInputSchema } from './blocks-schema';
+export {
+  localeSchema,
+  statusSchema,
+  documentIdSchema,
+  pageSchema,
+  pageSizeSchema,
+} from './input-schemas';
+export { buildDataSchema, buildComponentInputSchema, attributeToInputSchema } from './data-schema';
+export { buildSortSchema, getScalarAttributeKeys, SCALAR_ATTRIBUTE_TYPES } from './sort-schema';
+export { buildFiltersSchema, FILTER_OPERATORS, attributeTypeToFilterValue } from './filters-schema';
+export {
+  buildDocumentOutputSchema,
+  buildListOutputSchema,
+  buildDeleteOutputSchema,
+} from './output-schemas';
